Relocate RingBuffer testcases.
